await loadTimeSeriesData completion
