+++ README.md
@@ -16,12 +16,14 @@ See the [hyperobjekt core starter](https://gatsby-starter-hyperobjekt.netlify.ap
 
 ## Data Sources
 
-Data is pulled from a many sources:
+Data is pulled from many sources:
 
 - [UCLA Law COVID-19 Behind Bars Data Repo](https://github.com/uclalawcovid19behindbars/data)
 - [UCLA Law COVID-19 Behind Bars Google Sheet](https://docs.google.com/spreadsheets/d/1X6uJkXXS-O6eePLxw2e4JeRtM41uPZ2eRcOA_HkPVTk/edit#gid=1641553906)
+- [UCLA DOC Transparency Scorecard](https://docs.google.com/spreadsheets/d/1fHhRAjwYGVmgoHLUENvcYffHDjEQnpp7Rwt9tLeX_Xk/edit?gid=696812429)
+- [UCLA Time Series Data](http://104.131.72.50:3838/scraper_data/summary_data/scraped_time_series.csv)
 
-Data fetching and processing happens on build in `gatsby-node.js`
+Data fetching and processing happens on build in `gatsby-node.js`, with the exception of time series data, which is fetched and processed in the loadTimeSeriesData.js prebuild script. For local development, run `yarn prebuild` to generate the `/static/data/` CSVs necessary for the time series visuals. 
 
 ## Publishing the website
