+++ package.json
@@ -50,7 +50,7 @@
   ],
   "license": "MIT",
   "scripts": {
-    "prebuild": "node scripts/sourceTimeSeries.js",
+    "prebuild": "node scripts/loadTimeSeriesData.js",
     "build": "gatsby build",
     "develop": "gatsby develop -H 0.0.0.0",
     "clean": "gatsby clean",
